+++ NEWS.md
@@ -6,6 +6,9 @@
 - changed the default behaviour of estimateScaleTransforms() so that
 the default method for scatter channels is now "none" instead of 
 "linearQuantile"
+- changed default behaviour of ggplotEvents() and ggplotFilterEvents(), when
+logicle scale is used but no logicle parameters provided, these are now
+estimated using flowCore::estimatLogicle(), instead of explicit default values
 
 # CytoPipeline 1.1.1
 - tiny modifications to support upgrade to Bioc 3.18